Added nucleator attachments (working for sphere, not fully tested capsule) variables: STICK_TO_NUCLEATOR, RESTICK_TO_NUCLEATOR, NUC_LINK_FORCE, NUC_LINK_BREAKAGE_FORCE.  Attachment forces are plotted for bead and capsule (still needs proper scaling factor for capsule)  Added IMPOSED_NUC_ROT, IMPOSED_NUC_ROT_SPEED as testing to see how network responds to forcibly rotating nucleator (mainly to test nucleator attachement vector directions.)
Changed nucleation so that forces ramp up: repulsive force of harbinger start at zero and ramps linearly up to full after CROSSLINKDELAY iterations.  Also, changed nucleation so that new node cannot move if ALLOW_HARBINGERS_TO_MOVE is false, and added CROSSLINKDELAY flag to ini file.  Even if ALLOW_HARBINGERS_TO_MOVE is false, harbingers can repel one another.

Preliminary VISCOSITY and VISCOSITY_FACTOR for link-based viscosity (force proportional to delta-link-length).  DRAW_CAGE turns on drawing of nucleator dots.  

Added flags for WRITE_BMPS_PRE_SYMBREAK to draw all bitmaps from beginning even before symmetry breaking.

Fixed bug so that top cap of capsule force vectors are now plotted.
Changed scale so bead in now 5uM diameter.
Also changed some formatting so within 80 columns, added a 'Q' switch to turn off progress output for background running.  
Changed code to parse file so it reports unrecognized lines and output others on screen as they are read in.

+++ threadedtaskqueue.cpp
@@ -1,8 +1,9 @@
 #include "threadedtaskqueue.h"
 
+#include "stdafx.h"
 #include <assert.h>
-#include <iostream>
-// #include "stdafx.h"
+#include <iostream> 
+
 
 // static members are not *guaranteed* to have C linkage
 extern "C" void* wrapper_activate_workers(void *my_this)
@@ -108,7 +109,7 @@ void TaskQueue::start_tasks()
     finished_tasks = false;
     
     // set predicate and start the team
-    ntasks_todo = tasks.size(); // number in the queue
+    ntasks_todo = (int) tasks.size(); // number in the queue
     ready = true;
     int err = pthread_cond_broadcast(&start);
     if(err != 0){
